Work on simplifying the code
